fix: clear the clang-tidy backlog and gate on it
None of the fifteen findings was a bug, but leaving them advisory means the
next one that is a bug arrives in a list nobody reads. All fifteen are now
resolved, so GATE_CLANG_TIDY goes to true and any new finding is a regression
in the change that introduced it.

Fixed:

  bugprone-sizeof-expression -- static_assert(sizeof(ConfigCheck<Cfg>) > 0)
  existed only to force the config checks to instantiate. Client now derives
  privately from ConfigCheck<Cfg>: a base must be a complete type, so the
  assertions still fire, and an empty base costs nothing under EBO. Confirmed
  by measuring sizeof either way, and that a bad config still fails to build.

  readability-redundant-member-init x5 -- etl::delegate already default-
  constructs to unset.

  cppcoreguidelines-special-member-functions x3 -- Transport and Clock declare
  their copy and move operations protected and defaulted, the Core Guidelines
  shape for an abstract interface (C.67): a derived transport may still be
  copyable, but assigning through a Transport& can no longer slice it. Client
  deletes all four; it holds references, and moving one mid-session would
  leave the transport talking to a corpse.

  performance-enum-size -- Error was int16_t for twenty-three values. As
  uint8_t it packs better against the neighbouring members and takes eight
  bytes off every profile: Client<DefaultConfig> 4608 -> 4600. This is an
  API-visible change to a public enum, taken deliberately at 0.1.0, where the
  library is consumed from source, ships no install(EXPORT) and promises no
  ABI. One line to revert if that changes.

Declined and silenced in .clang-tidy with the reasoning next to each:
use-anyofallof would pull <algorithm> into a library that includes it nowhere;
vbi_size_c must stay one expression to work in a constant expression under
C++17; enqueue_sized invokes its callable once, in place.

clang-format stays advisory -- .clang-format still does not describe the tree,
and that needs a decision rather than a commit.

Verified: clang-tidy 0 findings, 94 tests pass under -Werror and under
ASan/UBSan, nm still shows no malloc/operator new/__cxa_throw/_Unwind_Resume,
and coverage holds at 87.9% line / 77.2% branch, above the 85/74 floor.

+++ include/mqtt/error.hpp
@@ -42,7 +42,11 @@ namespace mqtt {
 
 /// Every fallible operation in this library reports failure through this enum.
 /// There are no exceptions and no error globals.
-enum class Error : int16_t
+///
+/// Fixed at one byte. There are twenty-odd codes and room for two hundred more,
+/// and the narrower type packs better next to the other small members: it takes
+/// eight bytes off sizeof(Client<DefaultConfig>) on a 64-bit build for nothing.
+enum class Error : uint8_t
 {
     Ok = 0,
